Break the CLI tool into multiple files
