+++ package.json
@@ -10,10 +10,10 @@
   "scripts": {
     "docker:build": "docker build -t bbenligiray/hinter-cline:latest .",
     "docker:start": "docker run -it --rm -p8080:8080 -v./hinter-core-data:/app/hinter-core-data bbenligiray/hinter-cline:latest",
-    "start": "node hinter-helper.js"
+    "start": "node src/index.js"
   },
   "bin": {
-    "hinter-helper": "./hinter-helper.js"
+    "hinter-helper": "./src/index.js"
   },
   "dependencies": {
     "js-yaml": "^4.1.0"
